Fix 403 error on notification mark as read
- Added integer cast for user_id in Notification model
- Fixed type comparison in markAsRead method to handle int/string mismatch
- This fixes the "You do not have permission to access this notification" error

+++ app/Http/Controllers/NotificationController.php
@@ -89,8 +89,8 @@ class NotificationController extends Controller
      */
     public function markAsRead(Notification $notification)
     {
-        // Ensure user owns this notification
-        if ($notification->user_id !== Auth::id()) {
+        // Ensure user owns this notification (use loose comparison to handle int/string mismatch)
+        if ((int) $notification->user_id !== (int) Auth::id()) {
             abort(403, 'You do not have permission to access this notification.');
         }
